add thread-safe, checkable Queue
notifier will use the checkable queue to prevent
shipping the same exact exception instance twice.

previously, this might occur if sys.exc_info() is called
in the notifier on subsequent logs.

    ...
    except Exception:
        LOG.exception("An exception occurred.")

    LOG.error("Something else.") # gets the same traceback (bad!)

instead of taking the liberty and clearing exc_info, which
would have been the elegant, yet intrusive solution, the notifier
will just keep track of exception instances in a checkable,
thread safe queue, keeping track of the last 1000 exceptions logged.
The average exception instance size is about 70 Bytes, so the maximum
size this Queue should reach is less than 1/10 of a Megabyte.